+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ## Develop
 - Added hdf5 support for probes and added hdf5 I/O helper routines
 
+<<<<<<< feature/interfaceBC
+- Basic implementation of an overset interface boundary condition is added
+=======
 - Modify field_writer and probes to by default output in 
   `case.output_directory`
+>>>>>>> develop
 - Added MOST wall model and added diagnostics for the wall models.
 - Added the `data_streamer` simulation component, allowing data streaming 
   with ADIOS2.